split the test file
