+++ libcudacxx/include/cuda/std/__random/linear_congruential_engine.h
@@ -3,7 +3,7 @@
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES.
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES.
 //
 //===----------------------------------------------------------------------===//
 
@@ -34,81 +34,81 @@
 
 _CCCL_BEGIN_NAMESPACE_CUDA_STD
 
-template <uint64_t __A,
-          uint64_t __C,
-          uint64_t __M,
+template <uint64_t _Ap,
+          uint64_t _Cp,
           uint64_t _Mp,
-          bool _MightOverflow = (__A != 0 && __M != 0 && __M - 1 > (_Mp - __C) / __A),
-          bool _OverflowOk    = ((__M | (__M - 1)) > __M), // m = 2^n
-          bool _SchrageOk     = (__A != 0 && __M != 0 && __M % __A <= __M / __A)> // r <= q
+          uint64_t _MpMax,
+          bool _MightOverflow = (_Ap != 0 && _Mp != 0 && _Mp - 1 > (_MpMax - _Cp) / _Ap),
+          bool _OverflowOk    = ((_Mp | (_Mp - 1)) > _Mp), // m = 2^n
+          bool _SchrageOk     = (_Ap != 0 && _Mp != 0 && _Mp % _Ap <= _Mp / _Ap)> // r <= q
 struct __lce_alg_picker
 {
-  static_assert(__A != 0 || __M != 0 || !_MightOverflow || _OverflowOk || _SchrageOk,
+  static_assert(_Ap != 0 || _Mp != 0 || !_MightOverflow || _OverflowOk || _SchrageOk,
                 "The current values of a, c, and m cannot generate a number "
                 "within bounds of linear_congruential_engine.");
 
   static constexpr const bool __use_schrage = _MightOverflow && !_OverflowOk && _SchrageOk;
 };
 
-template <uint64_t __A,
-          uint64_t __C,
-          uint64_t __M,
+template <uint64_t _Ap,
+          uint64_t _Cp,
           uint64_t _Mp,
-          bool _UseSchrage = __lce_alg_picker<__A, __C, __M, _Mp>::__use_schrage>
+          uint64_t _MpMax,
+          bool _UseSchrage = __lce_alg_picker<_Ap, _Cp, _Mp, _MpMax>::__use_schrage>
 struct __lce_ta;
 
 // 64
 
-template <uint64_t __A, uint64_t __C, uint64_t __M>
-struct __lce_ta<__A, __C, __M, ~uint64_t{0}, true>
+template <uint64_t _Ap, uint64_t _Cp, uint64_t _Mp>
+struct __lce_ta<_Ap, _Cp, _Mp, ~uint64_t{0}, true>
 {
   using result_type = uint64_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
     // Schrage's algorithm
-    constexpr result_type __q = __M / __A;
-    constexpr result_type __r = __M % __A;
-    const result_type __t0    = __A * (__x % __q);
+    constexpr result_type __q = _Mp / _Ap;
+    constexpr result_type __r = _Mp % _Ap;
+    const result_type __t0    = _Ap * (__x % __q);
     const result_type __t1    = __r * (__x / __q);
-    __x                       = __t0 + (__t0 < __t1) * __M - __t1;
-    __x += __C - (__x >= __M - __C) * __M;
+    __x                       = __t0 + (__t0 < __t1) * _Mp - __t1;
+    __x += _Cp - (__x >= _Mp - _Cp) * _Mp;
     return __x;
   }
 };
 
-template <uint64_t __A, uint64_t __M>
-struct __lce_ta<__A, 0, __M, ~uint64_t{0}, true>
+template <uint64_t _Ap, uint64_t _Mp>
+struct __lce_ta<_Ap, 0, _Mp, ~uint64_t{0}, true>
 {
   using result_type = uint64_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
     // Schrage's algorithm
-    constexpr result_type __q = __M / __A;
-    constexpr result_type __r = __M % __A;
-    const result_type __t0    = __A * (__x % __q);
+    constexpr result_type __q = _Mp / _Ap;
+    constexpr result_type __r = _Mp % _Ap;
+    const result_type __t0    = _Ap * (__x % __q);
     const result_type __t1    = __r * (__x / __q);
-    __x                       = __t0 + (__t0 < __t1) * __M - __t1;
+    __x                       = __t0 + (__t0 < __t1) * _Mp - __t1;
     return __x;
   }
 };
 
-template <uint64_t __A, uint64_t __C, uint64_t __M>
-struct __lce_ta<__A, __C, __M, ~uint64_t{0}, false>
+template <uint64_t _Ap, uint64_t _Cp, uint64_t _Mp>
+struct __lce_ta<_Ap, _Cp, _Mp, ~uint64_t{0}, false>
 {
   using result_type = uint64_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    return (__A * __x + __C) % __M;
+    return (_Ap * __x + _Cp) % _Mp;
   }
 };
 
-template <uint64_t __A, uint64_t __C>
-struct __lce_ta<__A, __C, 0, ~uint64_t{0}, false>
+template <uint64_t _Ap, uint64_t _Cp>
+struct __lce_ta<_Ap, _Cp, 0, ~uint64_t{0}, false>
 {
   using result_type = uint64_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    return __A * __x + __C;
+    return _Ap * __x + _Cp;
   }
 };
 
@@ -120,16 +120,16 @@ struct __lce_ta<_Ap, _Cp, _Mp, ~uint32_t{0}, true>
   using result_type = uint32_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    constexpr auto __A = static_cast<result_type>(_Ap);
-    constexpr auto __C = static_cast<result_type>(_Cp);
-    constexpr auto __M = static_cast<result_type>(_Mp);
+    constexpr auto __a = static_cast<result_type>(_Ap);
+    constexpr auto __c = static_cast<result_type>(_Cp);
+    constexpr auto __m = static_cast<result_type>(_Mp);
     // Schrage's algorithm
-    constexpr result_type __q = __M / __A;
-    constexpr result_type __r = __M % __A;
-    const result_type __t0    = __A * (__x % __q);
+    constexpr result_type __q = __m / __a;
+    constexpr result_type __r = __m % __a;
+    const result_type __t0    = __a * (__x % __q);
     const result_type __t1    = __r * (__x / __q);
-    __x                       = __t0 + (__t0 < __t1) * __M - __t1;
-    __x += __C - (__x >= __M - __C) * __M;
+    __x                       = __t0 + (__t0 < __t1) * __m - __t1;
+    __x += __c - (__x >= __m - __c) * __m;
     return __x;
   }
 };
@@ -140,14 +140,14 @@ struct __lce_ta<_Ap, 0, _Mp, ~uint32_t{0}, true>
   using result_type = uint32_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    constexpr result_type __A = static_cast<result_type>(_Ap);
-    constexpr result_type __M = static_cast<result_type>(_Mp);
+    constexpr result_type __a = static_cast<result_type>(_Ap);
+    constexpr result_type __m = static_cast<result_type>(_Mp);
     // Schrage's algorithm
-    constexpr result_type __q = __M / __A;
-    constexpr result_type __r = __M % __A;
-    const result_type __t0    = __A * (__x % __q);
+    constexpr result_type __q = __m / __a;
+    constexpr result_type __r = __m % __a;
+    const result_type __t0    = __a * (__x % __q);
     const result_type __t1    = __r * (__x / __q);
-    __x                       = __t0 + (__t0 < __t1) * __M - __t1;
+    __x                       = __t0 + (__t0 < __t1) * __m - __t1;
     return __x;
   }
 };
@@ -158,10 +158,10 @@ struct __lce_ta<_Ap, _Cp, _Mp, ~uint32_t{0}, false>
   using result_type = uint32_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    constexpr result_type __A = static_cast<result_type>(_Ap);
-    constexpr result_type __C = static_cast<result_type>(_Cp);
-    constexpr result_type __M = static_cast<result_type>(_Mp);
-    return (__A * __x + __C) % __M;
+    constexpr result_type __a = static_cast<result_type>(_Ap);
+    constexpr result_type __c = static_cast<result_type>(_Cp);
+    constexpr result_type __m = static_cast<result_type>(_Mp);
+    return (__a * __x + __c) % __m;
   }
 };
 
@@ -171,32 +171,33 @@ struct __lce_ta<_Ap, _Cp, 0, ~uint32_t{0}, false>
   using result_type = uint32_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    constexpr result_type __A = static_cast<result_type>(_Ap);
-    constexpr result_type __C = static_cast<result_type>(_Cp);
-    return __A * __x + __C;
+    constexpr result_type __a = static_cast<result_type>(_Ap);
+    constexpr result_type __c = static_cast<result_type>(_Cp);
+    return __a * __x + __c;
   }
 };
 
 // 16
 
-template <uint64_t __A, uint64_t __C, uint64_t __M, bool __b>
-struct __lce_ta<__A, __C, __M, static_cast<uint16_t>(~0), __b>
+template <uint64_t _Ap, uint64_t _Cp, uint64_t _Mp, bool _UseSchrage>
+struct __lce_ta<_Ap, _Cp, _Mp, static_cast<uint16_t>(~0), _UseSchrage>
 {
   using result_type = uint16_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    return static_cast<result_type>(__lce_ta<__A, __C, __M, ~uint32_t{0}>::next(__x));
+    return static_cast<result_type>(__lce_ta<_Ap, _Cp, _Mp, ~uint32_t{0}>::next(__x));
   }
 };
 
 // 8
-template <uint64_t __A, uint64_t __C, uint64_t __M, bool __b>
-struct __lce_ta<__A, __C, __M, static_cast<uint8_t>(~0), __b>
+
+template <uint64_t _Ap, uint64_t _Cp, uint64_t _Mp, bool _UseSchrage>
+struct __lce_ta<_Ap, _Cp, _Mp, static_cast<uint8_t>(~0), _UseSchrage>
 {
   using result_type = uint8_t;
   [[nodiscard]] _CCCL_API static constexpr result_type next(result_type __x) noexcept
   {
-    return static_cast<result_type>(__lce_ta<__A, __C, __M, ~uint32_t{0}>::next(__x));
+    return static_cast<result_type>(__lce_ta<_Ap, _Cp, _Mp, ~uint32_t{0}>::next(__x));
   }
 };
